Extract some other utils for health score and version to common service
