+++ plugin-modernizer-core/src/main/java/io/jenkins/tools/pluginmodernizer/core/GuiceModule.java
@@ -6,7 +6,7 @@ import io.jenkins.tools.pluginmodernizer.core.github.GHService;
 import io.jenkins.tools.pluginmodernizer.core.impl.CacheManager;
 import io.jenkins.tools.pluginmodernizer.core.impl.PluginModernizer;
 import io.jenkins.tools.pluginmodernizer.core.utils.JdkFetcher;
-import io.jenkins.tools.pluginmodernizer.core.utils.UpdateCenterService;
+import io.jenkins.tools.pluginmodernizer.core.utils.PluginService;
 import org.apache.maven.shared.invoker.DefaultInvoker;
 import org.apache.maven.shared.invoker.Invoker;
 
@@ -23,7 +23,7 @@ public class GuiceModule extends AbstractModule {
         bind(Invoker.class).to(DefaultInvoker.class);
         bind(Config.class).toInstance(config);
         bind(CacheManager.class).toInstance(new CacheManager(config.getCachePath()));
-        bind(UpdateCenterService.class).toInstance(new UpdateCenterService());
+        bind(PluginService.class).toInstance(new PluginService());
         bind(GHService.class).toInstance(new GHService());
         bind(JdkFetcher.class).toInstance(new JdkFetcher(config.getCachePath()));
         bind(PluginModernizer.class).toInstance(new PluginModernizer());
